patchset bundle:
finish the tempfile size limit removal
blkio: symbolise the hard-coded block-size
Makefile: ensure correct permissions on install
paranoia: there are many places in the code which make unsafe assumptions
about implicit integer casts. There are a few fixes in this patchset,
but there is more work to be done.

+++ alloc.c
@@ -1,10 +1,10 @@
 #include "qed.h"
 
-void *qlloc(int);
-void *reqlloc(void *, int);
+void *qlloc(size_t);
+void *reqlloc(void *, size_t);
 
 void *
-qlloc(int n)
+qlloc(size_t n)
 {
    void *p;
    if((p=malloc(n))==NULL)
@@ -14,7 +14,7 @@ qlloc(int n)
 }
 
 void *
-reqlloc(void *p, int n)
+reqlloc(void *p, size_t n)
 {
    void *r;
    if((r=realloc(p,n))==NULL)
